+++ src/base/Routes.php
@@ -36,7 +36,7 @@ trait Routes
     protected function _registerCpRoutes(): void
     {
         Event::on(UrlManager::class, UrlManager::EVENT_REGISTER_CP_URL_RULES, function(RegisterUrlRulesEvent $event) {
-            $event->rules['contact-form-extensions'] = ['template' => 'contact-form-extensions/index'];
+            $event->rules['contact-form-extensions'] = 'contact-form-extensions/submissions/index';
             $event->rules['contact-form-extensions/submissions/<elementId:\\d+>'] = 'elements/edit';
             $event->rules['contact-form-extensions/submissions/<elementId:\\d+>/<siteHandle:{handle}>'] = 'elements/edit';
             $event->rules['contact-form-extensions/tools'] = 'contact-form-extensions/tools/index';